requirements.txt
# Core AWS
boto3>=1.34.0
botocore>=1.34.0

# LangChain (optional helper utilities)
langchain>=0.2.0
langchain-aws>=0.1.0

# Vector store + embeddings
faiss-cpu>=1.7.4
numpy>=1.26.0

# Streamlit UI
streamlit>=1.35.0

# HTTP + feeds
requests>=2.31.0
feedparser>=6.0.11

# PDF parsing
PyPDF2>=3.0.1

# Env management
python-dotenv>=1.0.0

# Utilities
# (uuid is a Python built-in – no separate install needed)
